Update and rename basicQT.c to quadtree.c
The quadtree now has a function to allow the program to determine whether points are inside a given range. This allows me to consult the quadtree in order to identify a point's closest neighbours. 

I created a player controlled avatar and attached the centre of the range to the avatar. Points which are determined to overlap with the range will appear blue. Points further away will appear red.