Married mapgen with quadtree and collision
Maps can now be generated based on array layout, translated onto the heap and the quadtree decides which walls are closest to the player and checks collision for them. 
Collision improved but far from perfect. 
I made the mistake of only inserting the walls' vectors into the quadtree, but I think it would be better to insert the entire struct instances into it, otherwise I lose easy access to a struct's members when I consult the quadtree. Worried about the impact on performance however and I am not going to do this until pathfinding has been understood.