Update pathfinding.c
Heavily annotated GPT script so that I can better understand. I adapted it to Raylib and made each cycle of the algorithm dependent upon spacebar input and coloured the nodes. This helped me to understand how the algorithm can be structured. 

Next steps will be further modification and marriage with already developed mechanics.